Updated address mapping
Updated the code that translates Amiga address to A314 addresses.

Currently the mapping is hardcoded for A314-A600. This means that a specific
a314.device needs to be compiled for A314-A600, with HARDCODED_A600_MAPPING = 1.

The A600 is assumed to have 2 MB chip memory, and all of that is A314 memory.